Got multiprocessing working
turned out the issue has something to do with the fork multiprocessing context. I got it working by switching to a spawn context, which starts a new python process for each gym environment, but the copying over of everything takes a relatively long time, so on my 4 core laptop I'm not seeing any improvement. I imagine that the time savings will scale with the number of CPU cores available. It'll also be influenced by the length of the episode; the longer the episode, the more the time savings compared to running batches sequentially. This should be really useful for the Atari implementation.